Demo1_1 support t3d::DDraw, t3d::DDSurface, t3d::ZBuffer, ...
git-svn-id: svn://127.0.0.1/repos/trunk@19 bc57d0f1-8121-e340-a336-9f9962c13ea2

+++ T3DLIB/t3dlib5.h
@@ -78,7 +78,7 @@ namespace t3d
 	class ConstSurfaceRef
 	{
 	public:
-		inline ConstSurfaceRef(elem_t * p, DWORD pitch)
+		inline ConstSurfaceRef(elem_t * p, LONG pitch)
 			: _p(p), _pitch(pitch)
 		{
 			assert(NULL != _p);
@@ -98,14 +98,14 @@ namespace t3d
 	protected:
 		elem_t * _p;
 
-		DWORD _pitch;
+		LONG _pitch;
 	};
 
 	template <typename elem_t>
 	class SurfaceRef : public ConstSurfaceRef<elem_t>
 	{
 	public:
-		inline SurfaceRef(elem_t * p, DWORD pitch)
+		inline SurfaceRef(elem_t * p, LONG pitch)
 			: ConstSurfaceRef<elem_t>(p, pitch)
 		{
 		}
